Split and refactor Multiple Filters Test
